Adds annotating and centering map on with searched items
This commit adds the ability to click on search results and annotate the
map with the selected result. Further annotations can be added and the
map will center to the most recent annotation whenever it is added.

If the users location change, the search region is adjusted (most of the
time, see quirks).

This continues to test my new-found understanding of architecture within
iOS, as I deal with both a LocationManager (for user location) and a
SearchService (for finding points of interest relative to region) in the
context of a single map view.

There are a few points where I grab data from these managers in
ContentView to adjust state and I don't think this is the right
approach. There's other update-y things happening in MapView when
annotations are changed but this doesn't feel "right"

Likely a blend of poor architecture and my current uncertainty in how to
structure this in iOS.

At the least, it works!

A few quirks yet:

* When the TextField appears for searching, it is not focused. This is
  driving me mad and is not possible in SwiftUI yet. Although it can be
  done with a blend of non-SwiftUI code and itself, I don't know if it's
  worth it for the sake of this sample
* The region is respected, sometimes! It might be an Xcode Simulator
  issue but sometimes I see "Phoenox, Arizona" in searches when I'm in
  Tokyo.